feat: Add support for custom schemes to Tracekit
Chromium Embedded Frameworks supports having custom schemes to load
content. This would cause tracekit to not match those stacktraces, as
the scheme was not in the fixed list that it searched for.

To prevent mismatches, the cause for "any scheme" looks ahead to ensure
that it is followed by a colon.